# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AI Code Mentor API
  version: 1.0.0
  description: API for generating AI-powered coding lessons and managing user profiles.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AppError:
      type: object
      properties:
        status:
          type: string
          example: fail
        error:
          type: string
          example: Validation Error
        details:
          type: array
          items:
            type: string

paths:
  /lessons/generate:
    post:
      summary: Generate a coding lesson
      description: Generates a contextual lesson using AI based on the curriculum week and day.
      tags:
        - Lessons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - semanaId
                - pomodoroIndex
              properties:
                semanaId:
                  type: integer
                  description: Week number (1-12)
                dia:
                  type: integer
                  description: Day number (1-7)
                diaIndex:
                  type: integer
                  description: Alternative to 'dia' (0-6)
                pomodoroIndex:
                  type: integer
                  description: Session index within the day
      responses:
        '200':
          description: Lesson generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  contenido:
                    type: string
                    description: HTML content of the lesson
                  metadata:
                    type: object
                    description: AI generation metadata
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'

  /profile:
    get:
      summary: Get user profile
      description: Retrieves the profile of the authenticated user or a public guest profile.
      tags:
        - Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                  capabilities:
                    type: array
                    items:
                      type: string
    post:
      summary: Update user profile
      description: Updates the profile of the authenticated user.
      tags:
        - Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
                preferences:
                  type: object
      responses:
        '200':
          description: Profile updated
        '401':
          description: Unauthorized